Observer pattern implementation to emit events from the SDK :dart:
