+++ shipengine_sdk/models/__init__.py
@@ -15,6 +15,7 @@ from .enums import (
     get_carrier_name_value,
 )
 from .package import (
+    EventOptions,
     Location,
     Package,
     Shipment,
